+++ src/api/cuda/Context.h
@@ -94,8 +94,8 @@ public:
     core::KernelLaunch &launch(core::Kernel &kernel);
     gmacError_t sync();
 
-    gmacError_t bufferToAccelerator(accptr_t dst, core::IOBuffer &buffer, size_t size, off_t off = 0);
-    gmacError_t acceleratorToBuffer(core::IOBuffer &buffer, const accptr_t dst, size_t size, off_t off = 0);
+    gmacError_t bufferToAccelerator(accptr_t dst, core::IOBuffer &buffer, size_t size, size_t off = 0);
+    gmacError_t acceleratorToBuffer(core::IOBuffer &buffer, const accptr_t dst, size_t size, size_t off = 0);
     gmacError_t waitAccelerator();
 
     void call(dim3 Dg, dim3 Db, size_t shared, cudaStream_t tokens);
